+++ src/main/java/no/fint/provider/student/model/SkoleressursRepository.java
@@ -2,6 +2,7 @@ package no.fint.provider.student.model;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.ImmutableSet;
+import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import no.fint.event.model.Event;
 import no.fint.event.model.ResponseStatus;
@@ -9,6 +10,7 @@ import no.fint.event.model.Status;
 import no.fint.model.resource.FintLinks;
 import no.fint.model.resource.utdanning.elev.SkoleressursResource;
 import no.fint.model.utdanning.elev.ElevActions;
+import no.fint.provider.student.behaviour.Behaviour;
 import no.fint.provider.student.service.Handler;
 import no.fint.provider.student.service.IdentifikatorFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,7 +36,11 @@ public class SkoleressursRepository implements Handler {
     @Autowired
     IdentifikatorFactory identifikatorFactory;
 
+    @Getter
     private Collection<SkoleressursResource> repository = new ConcurrentLinkedQueue<>();
+
+    @Autowired
+    private List<Behaviour<SkoleressursResource>> behaviours;
 
     @PostConstruct
     public void init() throws IOException {
@@ -61,10 +67,15 @@ public class SkoleressursRepository implements Handler {
                     List<SkoleressursResource> data = objectMapper.convertValue(response.getData(), objectMapper.getTypeFactory().constructCollectionType(List.class, SkoleressursResource.class));
                     log.trace("Converted data: {}", data);
                     data.stream().filter(i-> i.getSystemId()==null||i.getSystemId().getIdentifikatorverdi()==null).forEach(i->i.setSystemId(identifikatorFactory.create()));
-                    data.forEach(r -> repository.removeIf(i -> i.getSystemId().getIdentifikatorverdi().equals(r.getSystemId().getIdentifikatorverdi())));
                     repository.addAll(data);
                     response.setResponseStatus(ResponseStatus.ACCEPTED);
-                    response.setData(new ArrayList<>(data));
+                    response.setData(null);
+                    behaviours.forEach(b -> data.forEach(b.acceptPartially(response)));
+                    if (response.getResponseStatus() == ResponseStatus.ACCEPTED) {
+                        response.setData(new ArrayList<>(data));
+                        data.forEach(r -> repository.removeIf(i -> i.getSystemId().getIdentifikatorverdi().equals(r.getSystemId().getIdentifikatorverdi())));
+                        repository.addAll(data);
+                    }
                     break;
                 default:
                     response.setStatus(Status.ADAPTER_REJECTED);
